Move translation functions into separate file
