+++ src/type_ast/translate.rs
@@ -1,0 +1,353 @@
+use super::super::parse_ast as past;
+use super::nodes::*;
+use super::scope::{FuncScope, ModuleScope, Scope, ScopeLike};
+use super::typ::{Generic, Type, Variable};
+use super::{unify, Builtins, Closable, RecursionTracker, TypeError, TypeResult};
+
+pub fn translate_module(pmodule: past::Module) -> TypeResult<Module> {
+    let scope = ModuleScope::new().into_scope();
+
+    let mut statements = vec![];
+    for pstatement in pmodule.statements.into_iter() {
+        let statement = match pstatement {
+            past::ModuleStatement::Func(pfunc) => {
+                ModuleStatement::Func(translate_func(pfunc, scope.clone())?)
+            }
+            _ => unreachable!(),
+        };
+        statements.push(statement);
+    }
+
+    // Now that we've translated the whole module we can close all of the
+    // statements to ensure all types are bound.
+    let mut closed_statements = vec![];
+    for statement in statements.into_iter() {
+        closed_statements.push(statement.close(&mut RecursionTracker::new())?);
+    }
+
+    Ok(Module {
+        statements: closed_statements,
+    })
+}
+
+fn translate_func(pfunc: past::Func, scope: Scope) -> TypeResult<Func> {
+    let name = pfunc.name.name.clone();
+    // Build the `FuncArgument` nodes ahead of time so that they have types
+    // in place.
+    let arguments_nodes = pfunc
+        .arguments
+        .iter()
+        .map(|argument| FuncArgument {
+            name: argument.name.clone(),
+            // TODO: Support argument type definitions.
+            typ: Type::new_unbound(),
+        })
+        .collect::<Vec<_>>();
+
+    // We'll build the arguments and return types once and then reuse them for
+    // the recursive forward declaration and the real type.
+    let arguments = arguments_nodes
+        .iter()
+        .map(|argument| argument.typ.clone())
+        .collect::<Vec<_>>();
+    let retrn = Type::new_unbound();
+
+    let func_scope = FuncScope::new(Some(scope.clone())).into_scope();
+    for argument_node in arguments_nodes.iter() {
+        func_scope.add_local(&argument_node.name, argument_node.typ.clone())?;
+    }
+    // Build a forward declaration for the recursive call and add it to
+    // the function's scope.
+    let typ = Type::new_func(Some(name.clone()), arguments.clone(), retrn.clone());
+    func_scope.add_local(&name, typ.clone())?;
+
+    let body = match pfunc.body {
+        past::FuncBody::Block(block) => {
+            FuncBody::Block(translate_block(block, func_scope.clone())?)
+        }
+    };
+
+    let implicit_retrn = body.typ();
+    unify(&retrn, &implicit_retrn)?;
+
+    // Add the function to its defining scope.
+    scope.add_local(&name, typ.clone())?;
+
+    Ok(Func {
+        name: name.clone(),
+        arguments: arguments_nodes,
+        body,
+        scope: func_scope,
+        typ,
+    })
+}
+
+fn translate_block(pblock: past::Block, scope: Scope) -> TypeResult<Block> {
+    let mut statements = vec![];
+    for pstatement in pblock.statements {
+        let statement = match pstatement {
+            past::BlockStatement::CommentLine(_) => continue,
+            past::BlockStatement::Func(pfunc) => {
+                BlockStatement::Func(translate_func(pfunc, scope.clone())?)
+            }
+            past::BlockStatement::Expression(pexpression) => {
+                BlockStatement::Expression(translate_expression(&pexpression, scope.clone())?)
+            }
+        };
+        statements.push(statement);
+    }
+    let typ = if statements.is_empty() {
+        Type::new_empty_tuple()
+    } else {
+        statements.last().unwrap().typ()
+    };
+    Ok(Block {
+        statements,
+        span: pblock.span.clone(),
+        typ,
+    })
+}
+
+fn translate_expression(pexpression: &past::Expression, scope: Scope) -> TypeResult<Expression> {
+    Ok(match pexpression {
+        past::Expression::Identifier(pidentifier) => {
+            let name = pidentifier.name.clone();
+            let typ = scope.get_local(&name.name)?;
+            Expression::Identifier(Identifier { name, typ })
+        }
+        past::Expression::Infix(pinfix) => {
+            let lhs = translate_expression(&*pinfix.lhs, scope.clone())?;
+            let rhs = translate_expression(&*pinfix.rhs, scope)?;
+            // Left- and right-hand sides must be the same in an infix operation.
+            unify(lhs.typ(), rhs.typ())?;
+            let typ = rhs.typ().clone();
+            Expression::Infix(Infix {
+                lhs: Box::new(lhs),
+                op: pinfix.op.clone(),
+                rhs: Box::new(rhs),
+                typ,
+            })
+        }
+        past::Expression::LiteralInt(pliteral) => {
+            let class = Builtins::get("Int");
+            Expression::LiteralInt(LiteralInt {
+                value: pliteral.value,
+                typ: Type::new_object(class),
+            })
+        }
+        past::Expression::PostfixCall(pcall) => {
+            Expression::PostfixCall(translate_postfix_call(pcall, scope)?)
+        }
+        past::Expression::PostfixProperty(pproperty) => {
+            Expression::PostfixProperty(translate_postfix_property(pproperty, scope)?)
+        }
+        _ => unreachable!("Cannot translate {:?}", pexpression),
+    })
+}
+
+fn translate_postfix_call(pcall: &past::PostfixCall, scope: Scope) -> TypeResult<PostfixCall> {
+    let target = translate_expression(&*pcall.target, scope.clone())?;
+
+    let mut arguments = vec![];
+    for argument in pcall.arguments.iter() {
+        arguments.push(translate_expression(argument, scope.clone())?);
+    }
+
+    let retrn = {
+        // If the target is already a callable then unify directly instead of
+        // through a constraint.
+        if let Some((target_arguments, target_retrn)) = target.typ().maybe_callable() {
+            let call_arguments = arguments
+                .iter()
+                .map(|argument| argument.typ().clone())
+                .collect::<Vec<_>>();
+            if target_arguments.len() != call_arguments.len() {
+                return Err(TypeError::ArgumentLengthMismatch {
+                    expected: target_arguments,
+                    got: call_arguments,
+                });
+            }
+            for (target_argument, call_argument) in
+                target_arguments.iter().zip(call_arguments.iter())
+            {
+                unify(target_argument, call_argument)?;
+            }
+            // If the arguments unified then the return type is good.
+            target_retrn
+
+        // Otherwise build a callable generic constraint as an intermediary
+        // and unify through that.
+        } else {
+            // The return type of the callable.
+            let retrn = Type::new_unbound();
+
+            let mut generic = Generic::new();
+            generic.add_callable_constraint(
+                arguments
+                    .iter()
+                    .map(|argument| argument.typ().clone())
+                    .collect::<Vec<_>>(),
+                retrn.clone(),
+            );
+            // Unify to ensure target supports being called with the arguments and
+            // return types.
+            let intermediary = Type::new_variable(Variable::Generic(generic));
+            unify(target.typ(), &intermediary).map_err(|err| err.with_span(pcall.span.clone()))?;
+            retrn
+        }
+    };
+
+    Ok(PostfixCall {
+        target: Box::new(target),
+        arguments,
+        typ: retrn,
+    })
+}
+
+fn translate_postfix_property(
+    pproperty: &past::PostfixProperty,
+    scope: Scope,
+) -> TypeResult<PostfixProperty> {
+    let target = translate_expression(&*pproperty.target, scope.clone())?;
+
+    // The ultimate type of getting the target's property.
+    let typ = Type::new_unbound();
+
+    // Set up a variable generic with a property constraint.
+    let mut generic = Generic::new();
+    generic.add_property_constraint(pproperty.property.name.clone(), typ.clone());
+    // Apply unification to ensure the target supports having the
+    // given property.
+    let intermediary = Type::new_variable(Variable::Generic(generic));
+    unify(target.typ(), &intermediary).map_err(|err| err.with_span(pproperty.span.clone()))?;
+
+    Ok(PostfixProperty {
+        target: Box::new(target),
+        property: pproperty.property.clone(),
+        typ,
+    })
+}
+
+#[cfg(test)]
+mod tests {
+    use super::super::super::parse_ast as past;
+    use super::super::super::parser::{Location, Span, Token, Word};
+    use super::super::scope::{FuncScope, ScopeLike};
+    use super::super::{Builtins, Type, TypeError, Variable};
+    use super::{translate_expression, translate_func};
+
+    fn word<S: AsRef<str>>(name: S) -> Word {
+        Word {
+            name: name.as_ref().to_string(),
+            span: Span::unknown(),
+        }
+    }
+
+    #[test]
+    fn test_translate_infix() -> Result<(), TypeError> {
+        // Test simple addition of two integer literals.
+        let pexpression = past::Expression::Infix(past::Infix {
+            lhs: Box::new(past::Expression::LiteralInt(past::LiteralInt {
+                value: 1,
+                span: Span::unknown(),
+            })),
+            op: Token::Plus(Location::unknown()),
+            rhs: Box::new(past::Expression::LiteralInt(past::LiteralInt {
+                value: 2,
+                span: Span::unknown(),
+            })),
+        });
+        translate_expression(&pexpression, FuncScope::new(None).into_scope()).map(|_| ())?;
+
+        // Add an unbound variable with an integer literal; check that the
+        // variable is substituted to an integer.
+        let pexpression = past::Expression::Infix(past::Infix {
+            lhs: Box::new(past::Expression::Identifier(past::Identifier {
+                name: word("foo"),
+            })),
+            op: Token::Plus(Location::unknown()),
+            rhs: Box::new(past::Expression::LiteralInt(past::LiteralInt {
+                value: 2,
+                span: Span::unknown(),
+            })),
+        });
+        let foo = Type::new_unbound();
+        let scope = FuncScope::new(None).into_scope();
+        scope.add_local("foo", foo.clone())?;
+        translate_expression(&pexpression, scope).map(|_| ())?;
+        assert_eq!(
+            foo,
+            Type::new_substitute(Type::new_object(Builtins::get("Int")))
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_translate_func() -> Result<(), TypeError> {
+        let pfunc_inner = past::Func {
+            name: word("foo_inner"),
+            arguments: vec![],
+            body: past::FuncBody::Block(past::Block {
+                // statements: vec![past::BlockStatement::Expression(past::Expression::Infix(
+                //     past::Infix {
+                //         lhs: Box::new(past::Expression::Identifier(past::Identifier {
+                //             name: word("bar"),
+                //         })),
+                //         op: Token::Plus(Location::unknown()),
+                //         rhs: Box::new(past::Expression::LiteralInt(past::LiteralInt {
+                //             value: 1,
+                //             span: Span::unknown(),
+                //         })),
+                //     },
+                // ))],
+                statements: vec![past::BlockStatement::Expression(
+                    past::Expression::Identifier(past::Identifier { name: word("bar") }),
+                )],
+                span: Span::unknown(),
+            }),
+            span: Span::unknown(),
+        };
+
+        let pfunc = past::Func {
+            name: word("foo"),
+            arguments: vec![word("bar")],
+            body: past::FuncBody::Block(past::Block {
+                // statements: vec![past::BlockStatement::Expression(past::Expression::Infix(
+                //     past::Infix {
+                //         lhs: Box::new(past::Expression::Identifier(past::Identifier {
+                //             name: word("bar"),
+                //         })),
+                //         op: Token::Plus(Location::unknown()),
+                //         rhs: Box::new(past::Expression::LiteralInt(past::LiteralInt {
+                //             value: 1,
+                //             span: Span::unknown(),
+                //         })),
+                //     },
+                // ))],
+                statements: vec![
+                    past::BlockStatement::Func(pfunc_inner),
+                    past::BlockStatement::Expression(past::Expression::Infix(past::Infix {
+                        lhs: Box::new(past::Expression::Identifier(past::Identifier {
+                            name: word("bar"),
+                        })),
+                        op: Token::Plus(Location::unknown()),
+                        rhs: Box::new(past::Expression::LiteralInt(past::LiteralInt {
+                            value: 1,
+                            span: Span::unknown(),
+                        })),
+                    })),
+                ],
+                span: Span::unknown(),
+            }),
+            span: Span::unknown(),
+        };
+
+        let scope = FuncScope::new(None).into_scope();
+        let func = translate_func(pfunc, scope)?;
+
+        println!("{:?}", func);
+
+        Ok(())
+    }
+}
